remove repeated code and make a TextMixin
